Made View.java work
In the past when I tried, I couldn't use it.
