+++ lab-papers-please/java-classifcation/src/main/java/oop/practice/Main.java
@@ -75,16 +75,16 @@ public class Main {
     mapper.writeValue(new File("./lab-papers-please/output/marvel.json"), marvel);
 
     System.out.println("\n");
-    showStarWars();
+    View.showStarWars();
 
     System.out.println("\n");
-    showHitchhiker();
+    View.showHitchhiker();
 
     System.out.println("\n");
-    showMarvel();
+    View.showMarvel();
 
     System.out.println("\n");
-    showRings();
+    View.showRings();
   }
 
 
@@ -92,53 +92,4 @@ public class Main {
 record Universe(
     String name,
     List<JsonNode> individuals
-) { }
-public static void showStarWars() throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
-        File inputFile = new File("lab-papers-please\\output\\starwars.json");
-        JsonNode data = mapper.readTree(inputFile).get("individuals");
-        System.out.println("--------- Star Wars ---------");
-        for(JsonNode entry : data){
-
-            PrettyPrint.printEntry(entry);
-
-        }
-        System.out.println("-----------------------------");
-    }
-    public static void showHitchhiker() throws IOException{
-        ObjectMapper mapper = new ObjectMapper();
-        File inputFile = new File("lab-papers-please\\output\\hitchhiker.json");
-        JsonNode data = mapper.readTree(inputFile).get("individuals");
-        System.out.println("--------- Hitchhiker ---------");
-        for(JsonNode entry : data){
-
-            PrettyPrint.printEntry(entry);
-
-        }
-        System.out.println("------------------------------");
-    }
-    public static void showMarvel() throws IOException{
-        ObjectMapper mapper = new ObjectMapper();
-        File inputFile = new File("lab-papers-please\\output\\marvel.json");
-        JsonNode data = mapper.readTree(inputFile).get("individuals");
-        System.out.println("----------- Marvel -----------");
-        for(JsonNode entry : data){
-
-            PrettyPrint.printEntry(entry);
-
-        }
-        System.out.println("------------------------------");
-    }
-    public static void showRings() throws IOException{
-        ObjectMapper mapper = new ObjectMapper();
-        File inputFile = new File("lab-papers-please\\output\\rings.json");
-        JsonNode data = mapper.readTree(inputFile).get("individuals");
-        System.out.println("----- Lord of the Rings ------");
-        for(JsonNode entry : data){
-
-            PrettyPrint.printEntry(entry);
-
-        }
-        System.out.println("------------------------------");
-    }
-}
+) { }}
